refactor(metadata): migrate packagesigning component to v2
