+++ comp/metadata/packagesigning/def/component.go
