+++ comp/metadata/packagesigning/impl/readgpg.go
