+++ comp/metadata/packagesigning/impl/readgpg_test.go
