+++ comp/metadata/packagesigning/impl/rpmsigning.go
