+++ comp/metadata/packagesigning/impl/rpmsigning_test.go
