+++ comp/metadata/packagesigning/impl/testdata/debsig/policies/1A2B3C4D5E/dd-c.pol
