+++ comp/metadata/packagesigning/impl/testdata/yum.conf
